+++ NEWS.md
@@ -1,3 +1,10 @@
+## sps (under development)
+
+- Added `divisor_method()` to more easily generate divisor functions for
+`prop_allocation()` by name.
+
+- `prop_allocation()` is now faster.
+
 ## sps 0.6.1
 
 - Added a new vignette to show how inclusion probabilities are calculated.